Number of days per month in 1846-2007

Number of days in each month relevant to the Brest sea surge heights data BrestSurgeMaxima.

Usage

BrestSurgeDays

Format

BrestSurgeDays is a data frame with 162 rows (years 1846 to 2007) and the 12 variables (one for each month of the year). Each value in the data frame gives the number of days in the month in question.

The row names of BrestSurgeMaxima are the years 1946:2007 and the column names are the abbreviated names of the months.

See also

Examples

head(BrestSurgeDays)
#>      Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec
#> 1846  31  28  31  30  31  30  31  31  30  31  30  31
#> 1847  31  28  31  30  31  30  31  31  30  31  30  31
#> 1848  31  29  31  30  31  30  31  31  30  31  30  31
#> 1849  31  28  31  30  31  30  31  31  30  31  30  31
#> 1850  31  28  31  30  31  30  31  31  30  31  30  31
#> 1851  31  28  31  30  31  30  31  31  30  31  30  31