Methods for objects of class "confint_return_level"

Methods for objects of class "confint_return_level" returned from confint.return_level.

Usage

# S3 method for class 'confint_return_level'
print(x, ...)

# S3 method for class 'confint_return_level'
plot(x, parm = 1, add = TRUE, digits = 2, ...)

Arguments

x

An object inheriting from class "confint_return_level", a result of a call to confint.return_level.

...

Further arguments. For print.confint_return_level to pass arguments to print). For plot.confint_return_level to pass graphical parameters to plot to create the initial plot of the profile log-likelihood.

parm

An integer scalar. For which component, that is, which return level, in x we require a confidence interval.

add

A logical scalar. If add = TRUE then the plot is annotated with a horizontal line indicating the critical value for the profile log-likelihood used to calculate the confidence limits, vertical lines indicating the values of these limits and a legend stating the confidence interval.

digits

An integer. Passed to signif to round the confidence limits in the legend, if add = TRUE. The confidence level is hard-coded to be expressed to 3 significant figures.

Value

print.confint_return_level: the argument x is returned, invisibly.

plot.confint_return_level: a numeric vector containing the confidence interval for the return level chosen for the plot.

Details

print.confint_return_level. A numeric matrix with 2 columns giving the lower and upper confidence limits for the parameters specified by the argument parm in confint.return_level. These columns are labelled as (1-level)/2 and 1-(1-level)/2, expressed as a percentage, by default 2.5% and 97.5%.

plot.confint.return_level. A plot is produced of the profile log-likelihood for the parameter chosen by parm.

Examples

See return_level_methods.

See also